+++ fonctions_individu/decoder_indiv.h
@@ -3,6 +3,6 @@
 #include "../structures/individu.h"
 
 // Donne la valeur en base 10 de la personne grâce à sa séquence de bits.
-int decoder_indiv(individu);
+unsigned long long decoder_indiv(individu);
 
 #endif // DECODER_INDIV_H_INCLUDED
